+++ opentelemetry-sdk/tests/logs/test_export.py
@@ -46,7 +46,12 @@ from opentelemetry.sdk.environment_variables import (
 )
 from opentelemetry.sdk.resources import Resource as SDKResource
 from opentelemetry.sdk.util.instrumentation import InstrumentationScope
-from opentelemetry.trace import TraceFlags
+from opentelemetry.trace import (
+    NonRecordingSpan,
+    SpanContext,
+    TraceFlags,
+    set_span_in_context,
+)
 from opentelemetry.trace.span import INVALID_SPAN_CONTEXT
 
 EMPTY_LOG = LogData(
@@ -348,7 +353,7 @@ class TestBatchLogRecordProcessor(unittest.TestCase):
         logger.addHandler(LoggingHandler(logger_provider=provider))
 
         logger.error("error")
-        self.assertEqual(log_record_processor.emit.call_count, 1)
+        self.assertEqual(log_record_processor.on_emit.call_count, 1)
         log_record_processor.shutdown()
 
     def test_with_multiple_threads(self):  # pylint: disable=no-self-use
@@ -363,7 +368,7 @@ class TestBatchLogRecordProcessor(unittest.TestCase):
 
         def bulk_emit(num_emit):
             for _ in range(num_emit):
-                batch_processor.emit(EMPTY_LOG)
+                batch_processor.on_emit(EMPTY_LOG)
 
         total_expected_logs = 0
         with ThreadPoolExecutor(max_workers=69) as executor:
@@ -404,10 +409,10 @@ class TestBatchLogRecordProcessor(unittest.TestCase):
         # If `emit` calls logging.log then this test will throw a maximum recursion depth exceeded exception and fail.
         try:
             with self.assertNoLogs(sdk_logger, logging.NOTSET):
-                processor.emit(EMPTY_LOG)
+                processor.on_emit(EMPTY_LOG)
             processor.shutdown()
             with self.assertNoLogs(sdk_logger, logging.NOTSET):
-                processor.emit(EMPTY_LOG)
+                processor.on_emit(EMPTY_LOG)
         finally:
             sdk_logger.removeHandler(handler)
 
@@ -593,12 +598,20 @@ class TestBatchLogRecordProcessor(unittest.TestCase):
 class TestConsoleLogExporter(unittest.TestCase):
     def test_export(self):  # pylint: disable=no-self-use
         """Check that the console exporter prints log records."""
+        ctx = set_span_in_context(
+            NonRecordingSpan(
+                SpanContext(
+                    2604504634922341076776623263868986797,
+                    5213367945872657620,
+                    False,
+                    TraceFlags(0x01),
+                )
+            )
+        )
         log_data = LogData(
             log_record=LogRecord(
                 timestamp=int(time.time() * 1e9),
-                trace_id=2604504634922341076776623263868986797,
-                span_id=5213367945872657620,
-                trace_flags=TraceFlags(0x01),
+                context=ctx,
                 severity_text="WARN",
                 severity_number=SeverityNumber.WARN,
                 body="Zhengzhou, We have a heaviest rains in 1000 years",