+++ CHANGELOG.md
@@ -26,8 +26,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#4637](https://github.com/open-telemetry/opentelemetry-python/pull/4637))
 - Logging API accepts optional `context`; deprecates `trace_id`, `span_id`, `trace_flags`.
   ([#4597](https://github.com/open-telemetry/opentelemetry-python/pull/4597))
+- sdk: use context instead of trace_id,span_id for initializing LogRecord
+  ([#4653](https://github.com/open-telemetry/opentelemetry-python/pull/4653))
+- Rename LogRecordProcessor.emit to on_emit
+  ([#4648](https://github.com/open-telemetry/opentelemetry-python/pull/4648))
 - Logging API hide std_to_otel function to convert python logging severity to otel severity
   ([#4649](https://github.com/open-telemetry/opentelemetry-python/pull/4649))
+- proto: relax protobuf version requirement to support v6
+  ([#4620](https://github.com/open-telemetry/opentelemetry-python/pull/4620))
 
 ## Version 1.34.0/0.55b0 (2025-06-04)